more SP optimizations; fixed initial data sql insert; fixed group management bug

+++ PhpbbInDotnet.Objects/UpsertGroupDto.cs
@@ -14,7 +14,7 @@
 
         public string? DbColor => Color?.TrimStart('#');
 
-        public int UploadLimit { get; set; }
+        public long UploadLimit { get; set; }
 
         public int EditTime { get; set; }
